Updated SRC
Updated the full source code. With a direct zip file containing the full source code to the Chrome extension. source code prebuilt to only need to use the filter options enabled to filter options. side load the unpacked extension and then you're ready to go. Currently you still working on integrating the server side? endpoint Functionality, I already have it set up to work. It connects to the server. But I think the problem is that I already have almost 600,000 domains in a block list. Chrome is giving me issues where a limitation of I'm going over the limit of how many domains I can be blocking at any given time. So therefore this is still a work in progress. And further development. Will be ongoing.